:truck: separate batch correction from normalization (#46)
* :truck: separate batch correction from normalization

- fix Move batch correction outside Normalisation 💄
Fixes #22

* :bug: sync normalization nb and update test imports

* :art: format and remove unused import

* :bug: sync nb

* :bug: add batch correction to website index

* :art: link module api
